Rename StorageClassForm/Table to CopyOptionsForm/Table
