+++ src/app/plan/components/Wizard/WizardComponent.tsx
@@ -3,7 +3,7 @@ import { Wizard } from '@patternfly/react-core';
 import GeneralForm from './GeneralForm';
 import ResourceSelectForm from './ResourceSelectForm';
 import VolumesForm from './VolumesForm';
-import StorageClassForm from './StorageClassForm';
+import CopyOptionsForm from './CopyOptionsForm';
 import ResultsStep from './ResultsStep';
 import { PollingContext } from '../../../home/duck/context';
 import { FormikProps } from 'formik';
@@ -163,7 +163,7 @@ const WizardComponent = (props: IOtherProps & FormikProps<IFormValues>) => {
           name: 'Copy options',
           component: (
             <WizardStepContainer title="Copy options">
-              <StorageClassForm
+              <CopyOptionsForm
                 values={values}
                 setFieldValue={setFieldValue}
                 currentPlan={currentPlan}
